+++ packages/agw-client/src/utils.ts
@@ -15,7 +15,7 @@ import {
   type UnionRequiredBy,
 } from 'viem';
 import { parseAccount } from 'viem/accounts';
-import { abstractTestnet } from 'viem/chains';
+import { abstract, abstractTestnet } from 'viem/chains';
 import {
   type ChainEIP712,
   type SignEip712TransactionParameters,
@@ -33,7 +33,7 @@ import { type Call } from './types/call.js';
 // TODO: support Abstract mainnet
 export const VALID_CHAINS: Record<number, Chain> = {
   [abstractTestnet.id]: abstractTestnet,
-  [2741]: { id: 2741 } as Chain,
+  [abstract.id]: abstract,
 };
 
 export function convertBigIntToString(value: any): any {
@@ -169,32 +169,39 @@ export function transformEip712TypedData(
   }
 
   return {
-    chainId: typedData.domain.chainId,
+    chainId: Number(typedData.domain.chainId),
     account: parseAccount(
-      toHex(typedData.message['from'] as bigint, {
+      toHex(BigInt(typedData.message['from'] as string), {
         size: 20,
       }),
     ),
-    to: toHex(typedData.message['to'] as bigint, {
+    to: toHex(BigInt(typedData.message['to'] as string), {
       size: 20,
     }),
-    gas: typedData.message['gasLimit'] as bigint,
-    gasPerPubdata: typedData.message['gasPerPubdataByteLimit'] as bigint,
-    maxFeePerGas: typedData.message['maxFeePerGas'] as bigint,
-    maxPriorityFeePerGas: typedData.message['maxPriorityFeePerGas'] as bigint,
+    gas: BigInt(typedData.message['gasLimit'] as string),
+    gasPerPubdata: BigInt(
+      typedData.message['gasPerPubdataByteLimit'] as string,
+    ),
+    maxFeePerGas: BigInt(typedData.message['maxFeePerGas'] as string),
+    maxPriorityFeePerGas: BigInt(
+      typedData.message['maxPriorityFeePerGas'] as string,
+    ),
     paymaster:
-      (typedData.message['paymaster'] as bigint) > 0n
-        ? toHex(typedData.message['paymaster'] as bigint, {
+      (typedData.message['paymaster'] as string) != '0'
+        ? toHex(BigInt(typedData.message['paymaster'] as string), {
             size: 20,
           })
         : undefined,
     nonce: typedData.message['nonce'] as number,
-    value: typedData.message['value'] as bigint,
-    data: typedData.message['data'] as Hex,
+    value: BigInt(typedData.message['value'] as string),
+    data:
+      typedData.message['data'] === '0x0'
+        ? '0x'
+        : (typedData.message['data'] as Hex),
     factoryDeps: typedData.message['factoryDeps'] as Hex[],
     paymasterInput:
-      typedData.message['paymasterParams'] !== '0x'
-        ? (typedData.message['paymasterParams'] as Hex)
+      typedData.message['paymasterInput'] !== '0x'
+        ? (typedData.message['paymasterInput'] as Hex)
         : undefined,
   };
 }