+++ packages/agw-client/test/src/walletActions.test.ts
@@ -92,6 +92,10 @@ describe('globalWalletActions', () => {
       mockSignerClient,
       mockArgs,
       EOA_VALIDATOR_ADDRESS,
+      false,
+      {},
+      undefined,
+      false,
     );
   });
 
@@ -122,6 +126,7 @@ describe('globalWalletActions', () => {
       mockSignerClient,
       mockPublicClient,
       mockArgs,
+      false,
     );
   });