+++ packages/web3-react-agw/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @abstract-foundation/web3-react-agw
 
+## 1.3.0
+
+### Minor Changes
+
+- 7de8115: Add non executing transaction signing
+
+### Patch Changes
+
+- Updated dependencies [7de8115]
+- Updated dependencies [97015f9]
+  - @abstract-foundation/agw-client@1.3.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 3cf7f67: Update provider config; require chain definition
+
+### Patch Changes
+
+- Updated dependencies [3cf7f67]
+  - @abstract-foundation/agw-client@1.2.0
+
 ## 1.1.1
 
 ### Patch Changes